Feat : 공고 디테일 모달 연결
